# openapi/devflow-lengkap.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevFlow — API Lengkap",
    "version": "1.0",
    "description": "Gabungan spesifikasi kelima service backend DevFlow, seluruhnya di satu berkas.\n\nDihasilkan dari controller lewat springdoc, lalu digabung — bukan ditulis tangan, jadi isinya mengikuti kode.\n\nSetiap operasi diberi tag nama service asalnya. Ambil token lewat POST /api/v1/auth/login, lalu pakai skema bearerAuth."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "api-gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/workspaces/{id}/details": {
      "put": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "updateWorkspaceDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDetailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "getUserWorkspaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "createWorkspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{id}/invitations": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "getWorkspaceInvitations",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "inviteMember",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/projects/{projectId}/invitations": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "getProjectInvitations",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "inviteProjectMember",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/{token}/accept": {
      "post": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "acceptInvitation",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/workspaces/{id}": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "getWorkspaceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{id}/members": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "getWorkspaceMembers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceMemberDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/projects/{projectId}/members": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "getProjectMembers",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceMemberDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/check-slug": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "checkSlug",
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/{token}": {
      "get": {
        "tags": [
          "workspace-service"
        ],
        "operationId": "getInvitationInfo",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/integrations/github/workspaces/{workspaceId}": {
      "get": {
        "tags": [
          "integration-service"
        ],
        "operationId": "getInstallations",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GithubInstallation"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateWorkspaceDetailsRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "businessType": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        }
      },
      "WorkspaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "businessType": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "ARCHIVED"
            ]
          },
          "userRole": {
            "type": "string",
            "enum": [
              "OWNER",
              "DEVELOPER",
              "CLIENT"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2
          },
          "logoUrl": {
            "type": "string"
          },
          "businessType": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "InviteMemberRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "DEVELOPER",
              "CLIENT"
            ]
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "DEVELOPER",
              "CLIENT"
            ]
          },
          "token": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "EXPIRED",
              "REVOKED"
            ]
          },
          "invitedBy": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WorkspaceMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GithubInstallation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "installationId": {
            "type": "string"
          },
          "accountLogin": {
            "type": "string"
          },
          "accountType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "installedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Access token dari /api/v1/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}