# workspace-service/src/main/java/com/saas/workspace_service/dto/CreateWorkspaceRequest.java
package com.saas.workspace_service.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class CreateWorkspaceRequest {
    @NotBlank
    @Size(min = 2, max = 100)
    private String name;

    @NotBlank
    @Size(min = 2, max = 100)
    private String slug;

    private String logoUrl;
    private String businessType;
    private String timezone;
}